requirements.txt
huggingface_hub
transformers
datasets
evaluate
torch
pandas
tqdm
PyYAML
